# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "McpServerStatus": {
        "properties": {
          "asOf": {
            "format": "date-time",
            "title": "Asof",
            "type": "string"
          },
          "autostart": {
            "title": "Autostart",
            "type": "boolean"
          },
          "commandLine": {
            "title": "Commandline",
            "type": "string"
          },
          "cwd": {
            "title": "Cwd",
            "type": "string"
          },
          "exitCode": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exitcode"
          },
          "key": {
            "title": "Key",
            "type": "string"
          },
          "lastError": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lasterror"
          },
          "logSource": {
            "title": "Logsource",
            "type": "string"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "port": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PortInfo"
              },
              {
                "type": "null"
              }
            ]
          },
          "recentOutput": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Recentoutput",
            "type": "array"
          },
          "restartOnCrash": {
            "title": "Restartoncrash",
            "type": "boolean"
          },
          "restarts": {
            "default": 0,
            "title": "Restarts",
            "type": "integer"
          },
          "revision": {
            "title": "Revision",
            "type": "integer"
          },
          "startedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Startedat"
          },
          "state": {
            "enum": [
              "stopped",
              "starting",
              "running",
              "stopping",
              "crashed"
            ],
            "title": "State",
            "type": "string"
          },
          "stoppedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stoppedat"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "tools": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/McpToolList"
              },
              {
                "type": "null"
              }
            ]
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          }
        },
        "required": [
          "revision",
          "asOf",
          "key",
          "title",
          "url",
          "commandLine",
          "cwd",
          "autostart",
          "restartOnCrash",
          "state",
          "logSource"
        ],
        "title": "McpServerStatus",
        "type": "object"
      },
      "McpTool": {
        "properties": {
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "params": {
            "items": {
              "$ref": "#/components/schemas/McpToolParam"
            },
            "title": "Params",
            "type": "array"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          }
        },
        "required": [
          "name",
          "params"
        ],
        "title": "McpTool",
        "type": "object"
      },
      "McpToolList": {
        "properties": {
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "fetchedAt": {
            "format": "date-time",
            "title": "Fetchedat",
            "type": "string"
          },
          "protocolVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Protocolversion"
          },
          "serverName": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Servername"
          },
          "serverVersion": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Serverversion"
          },
          "tools": {
            "items": {
              "$ref": "#/components/schemas/McpTool"
            },
            "title": "Tools",
            "type": "array"
          }
        },
        "required": [
          "fetchedAt",
          "tools"
        ],
        "title": "McpToolList",
        "type": "object"
      },
      "McpToolParam": {
        "properties": {
          "default": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Default"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "required": {
            "title": "Required",
            "type": "boolean"
          },
          "type": {
            "title": "Type",
            "type": "string"
          }
        },
        "required": [
          "name",
          "type",
          "required"
        ],
        "title": "McpToolParam",
        "type": "object"
      },
...
      "PortInfo": {
        "properties": {
          "checkedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Checkedat"
          },
          "managed": {
            "default": false,
            "title": "Managed",
            "type": "boolean"
          },
          "open": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Open"
          },
          "pid": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pid"
          },
          "port": {
            "title": "Port",
            "type": "integer"
          },
          "process": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Process"
          }
        },
        "required": [
          "port",
          "open"
        ],
        "title": "PortInfo",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "AI Control Center",
    "version": "0.8.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/mcp/servers/{key}": {
      "get": {
        "operationId": "mcp_server",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Server",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/servers/{key}/restart": {
      "post": {
        "operationId": "mcp_restart",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "stopped, but the new start failed"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Restart",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/servers/{key}/start": {
      "post": {
        "description": "Returns right after the process exists (state starting); readiness follows via SSE mcp.server.",
        "operationId": "mcp_start",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "cannot start: program/folder missing or port taken"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/servers/{key}/stop": {
      "post": {
        "description": "Waits until the whole process tree has ended (at most stop_timeout_s + a few seconds).",
        "operationId": "mcp_stop",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpServerStatus"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stop",
        "tags": [
          "mcp"
        ]
      }
    },
    "/api/v1/mcp/servers/{key}/tools": {
      "get": {
        "description": "Asks the server now (MCP tools/list). Works for a server started elsewhere too, if the URL answers.\n\nFailures are part of the answer (error field), not an HTTP error: an unreachable server is a state.",
        "operationId": "mcp_tools",
        "parameters": [
          {
            "in": "path",
            "name": "key",
            "required": true,
            "schema": {
              "title": "Key",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/McpToolList"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "description": "server has no url configured"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Tools",
        "tags": [
          "mcp"
        ]
      }
    },
...
  }
}